Add files via upload
Spoke to U about adding in one of my portfolio websites as a backup option. This one was created for the purpose of selling nfts, i gutted out most of the nft stuff and left the index to show a conceptual idea of what the website could look like if we end up going that route.
Added in the name and changed some h1 tags to help visuallise it.

Also created a new BB logo cuz I did not like the circular design in webform. I think the new one is much more modern looking. Can send over the svg files if we wanna keep this logo